requirements.txt
colorlog==4.7.2
torchvision==0.8.2
tqdm==4.47.0
torch==1.7.1
munch==2.5.0
PyYAML==5.3.1
tensorboardX==2.1
